design/Analysis-Platform-main 4/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg==3.1.12
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-extra-types==2.4.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==3.2.2
pandas==2.1.4
joblib==1.3.2
scikit-learn==1.3.2
xgboost==2.0.3
python-multipart==0.0.9
email-validator==2.1.1
requests
